certain endpoints are only made for admin to use (#12)
* package fixes

* admin exception creation

* pom xml modify

* removing unused method

* error messages class to messages

* is admin method implementation

* merge bug fixes

* user response and admin response implementation

* only admins can promote users to admin

* modify informations endpoints creation

* modify email in resource creation

* using user response for get user by id

* modify phone in resource creation

* modify name in resource creation

* modify surname in resource creation

* modify password in resource creation

* send email verification message fix

* modify name bug fix

* modify email request creation

* modify phone request creation

* modify name request creation

* modify surname request creation

* modify password request creation

* using session id to check user information to update their names

* implementing update methods for user

* validate email implementation

* modify surname implementation

* modify password request

* password update method implementation

* fixing promoting user method

* email update method implementation fix and  bug fixing get user by id

* only admin can certain things in talk class

* only admin can do certain things in tag class

* only admin can do certain things in speaker class

* only admin can do certain things in partner class

* only admin can do certain things in event class

---------

+++ src/main/java/fullstack/service/EventService.java
@@ -1,8 +1,11 @@
 package fullstack.service;
 
 import fullstack.persistence.repository.EventRepository;
+import fullstack.service.exception.AdminAccessException;
+import fullstack.service.exception.UserNotFoundException;
 import io.quarkus.hibernate.orm.panache.PanacheRepository;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
 import fullstack.persistence.model.Event;
 
@@ -13,9 +16,12 @@ import java.util.UUID;
 @ApplicationScoped
 public class EventService implements PanacheRepository<Event> {
     private final EventRepository eventRepository;
+    private final UserService userService;
 
-    public EventService(EventRepository eventRepository) {
+    @Inject
+    public EventService(EventRepository eventRepository, UserService userService) {
         this.eventRepository = eventRepository;
+        this.userService = userService;
     }
 
     public List<Event> getAllEvents() {
@@ -39,20 +45,29 @@ public class EventService implements PanacheRepository<Event> {
     }
 
     @Transactional
-    public Event save(Event event) {
+    public Event save(String sessionId, Event event) throws UserNotFoundException {
+        if (userService.isAdmin(sessionId)) {
+            throw new AdminAccessException("Accesso negato. Solo gli amministratori possono promuovere altri utenti ad admin.");
+        }
         event.setId(UUID.randomUUID().toString());
         persist(event);
         return event;
     }
 
     @Transactional
-    public void deleteById(String id) {
+    public void deleteById(String sessionId, String id) throws UserNotFoundException {
+        if (userService.isAdmin(sessionId)) {
+            throw new AdminAccessException("Accesso negato. Solo gli amministratori possono promuovere altri utenti ad admin.");
+        }
         eventRepository.deleteById(id);
     }
 
-        @Transactional
-    public int update(String id, Event event) {
-            return eventRepository.update(id, event);
+    @Transactional
+    public int update(String sessionId, String id, Event event) throws UserNotFoundException {
+        if (userService.isAdmin(sessionId)) {
+            throw new AdminAccessException("Accesso negato. Solo gli amministratori possono promuovere altri utenti ad admin.");
+        }
+        return eventRepository.update(id, event);
     }
 }
