+++ src/main/java/fullstack/service/PartnerService.java
@@ -1,6 +1,8 @@
 package fullstack.service;
 
 import fullstack.persistence.repository.PartnerRepository;
+import fullstack.service.exception.AdminAccessException;
+import fullstack.service.exception.UserNotFoundException;
 import io.quarkus.hibernate.orm.panache.PanacheRepository;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -12,8 +14,14 @@ import java.util.UUID;
 
 @ApplicationScoped
 public class PartnerService {
+
+    private final PartnerRepository partnerRepository;
+    private final UserService userService;
     @Inject
-    PartnerRepository partnerRepository;
+    public PartnerService(PartnerRepository partnerRepository, UserService userService) {
+        this.partnerRepository = partnerRepository;
+        this.userService = userService;
+    }
 
     public List<Partner> getAllPartners() {
         return partnerRepository.listAll();
@@ -24,19 +32,28 @@ public class PartnerService {
     }
 
     @Transactional
-    public Partner save(Partner partner) {
+    public Partner save(String sessionId, Partner partner) throws UserNotFoundException {
+        if (userService.isAdmin(sessionId)) {
+            throw new AdminAccessException("Accesso negato. Solo gli amministratori possono promuovere altri utenti ad admin.");
+        }
         partner.setId(UUID.randomUUID().toString());
         partnerRepository.persist(partner);
         return partner;
     }
 
     @Transactional
-    public void deleteById(String id) {
+    public void deleteById(String sessionId, String id) throws UserNotFoundException {
+        if (userService.isAdmin(sessionId)) {
+            throw new AdminAccessException("Accesso negato. Solo gli amministratori possono promuovere altri utenti ad admin.");
+        }
         partnerRepository.deleteById(id);
     }
 
     @Transactional
-    public int update(String id, Partner partner) {
+    public int update(String sessionId, String id, Partner partner) throws UserNotFoundException {
+        if (userService.isAdmin(sessionId)) {
+            throw new AdminAccessException("Accesso negato. Solo gli amministratori possono promuovere altri utenti ad admin.");
+        }
         return partnerRepository.update(id, partner);
     }
